requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
scikit-learn>=1.4.0
numpy>=1.26.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
nltk>=3.8.0
pydantic>=2.0.0
